refactor: extract LauncherModeKeycap from OverlayKeycapView
Move launcher mode rendering (app icons, favicons, link badges,
transition animation, fallback symbols) into a focused child view
with 11 parameters. Parent dispatches to it based on mode. Old
extension file emptied.

Snapshot tests pass pixel-identical. Phase 4.2 of the overlay
keycap decomposition plan.
